Divide the manage roles page into appropriate sub components
