+++ client/src/components/App.js
@@ -7,7 +7,7 @@ import PageInternalError from "./general/PageInternalError";
 import PageNotFound from "./general/PageNotFound";
 import Home from "./Home";
 import Login from "./Login";
-import SetRoles from "./set_roles/SetRoles";
+import ManageRoles from "./set_roles/ManageRoles";
 
 const globalStyles = css`
   @import url('https://fonts.googleapis.com/css?family=Muli');
@@ -119,7 +119,7 @@ function App() {
           <StudentCreatePlan />
         </Route>
         <Route path="/manageRoles">
-          <SetRoles />
+          <ManageRoles />
         </Route>
         <Route path="/500">
           <PageInternalError />
